fix menu overflowing
