+++ src/IbdTracker.Frontend/src/pages/dashboard/_layout.svelte
@@ -2,7 +2,10 @@
     import { ready } from "@roxi/routify";
     import { isLoading, isAuthenticated } from "../../stores/authStore";
     import Login from "../login/index.svelte";
-    import Navigation from "../../components/navigation/Navigation.svelte";
+    import Menu from "../../components/navigation/Menu.svelte";
+    import Header from "../../components/navigation/Headbar.svelte";
+
+    import { menuOpened } from "../../stores/menuStore";
 
     $ready();
 </script>
@@ -14,11 +17,14 @@
         <div />
     {:else if $isAuthenticated}
         <div class="lg:flex w-full">
-            <Navigation />
-            <div class="mt-8 mx-6 lg:mx-20 max-w-2xl lg:w-full">
-                <slot />
+            <Menu />
+            <div class:max-h-screen={$menuOpened} class:overflow-hidden={$menuOpened}>
+                <Header />
+                <div class="mt-8 mx-6 lg:mx-20 max-w-2xl lg:w-full">
+                    <slot />
+                </div>
             </div>
-        </div>        
+        </div>
     {:else}
         <Login />
     {/if}
