+++ src/IbdTracker.Frontend/src/stores/menuStore.ts
@@ -1,0 +1,3 @@
+import { Writable, writable } from "svelte/store";
+
+export let menuOpened: Writable<boolean> = writable(false);
